category section managed with some bug fix

+++ app/Http/Controllers/Admin/AdminCategoryController.php
@@ -5,34 +5,40 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Inertia\Inertia;
 
 class AdminCategoryController extends Controller
 {
-    public function categoryPage()
+    public function categoriesPage()
     {
-        $categories = Category::withCount('products')->get();
+        $categories = Category::withCount('products')->orderByDesc('id')->paginate(10);
         return inertia('Backend/Categories', ['categories' => $categories]);
+    }
+
+    public function createCategoryPage(Request $request)
+    {
+        return Inertia::render('Backend/AddCategory');
     }
     public function createCategory(Request $request)
     {
-        $validatedRequest = $request->validate([
+        $request->validate([
             'name' => 'required|max:30',
             'slug' => 'required|alpha_dash|unique:categories,slug',
             'description' => 'required',
             'image' => 'required|image|max:2048',
             'status' => 'boolean',
             'meta_title' => 'required|max:30',
-            'meta_desc' => 'required|max:100',
+            'meta_description' => 'required|max:100',
         ]);
 
-        if ($validatedRequest) {
+        try {
             $category = new Category;
             $category->name = $request->name;
             $category->slug = $request->slug;
             $category->description = $request->description;
-            $category->status = $request->status == null ? 0 : 1;
+            $category->status = $request->status;
             $category->meta_title = $request->meta_title;
-            $category->meta_description = $request->meta_desc;
+            $category->meta_description = $request->meta_description;
 
             $image = $request->file('image');
             $imageName = "category" . '-' . $request->slug . '.' . $image->getClientOriginalExtension();
@@ -41,37 +47,38 @@ class AdminCategoryController extends Controller
             $category->save();
 
             return redirect()->back();
+        } catch (\Exception $e) {
+            return back()->withErrors(['message' => 'Something Wrong!']);
         }
-
-        return redirect()->back();
     }
 
     public function editCategoryPage($id)
     {
         $category = Category::find($id);
-        return view('components.back-end.edit-category', compact('category'));
+        return inertia('Backend/EditCategory', ['category' => $category]);
     }
 
-    public function editCategory(Request $request)
+    public function editCategory(Request $request, $id)
     {
-        $validatedRequest = $request->validate([
-            'id' => 'required|exists:categories,id',
+        $request->validate([
             'name' => 'required|max:30',
             'slug' => 'required|alpha_dash|unique:categories,slug,' . $request->id,
             'description' => 'required',
-            'image' => 'image|max:2048',
+            'image' => 'nullable|image|max:2048',
             'status' => 'boolean',
             'meta_title' => 'required|max:30',
-            'meta_desc' => 'required|max:100',
+            'meta_description' => 'required|max:100',
         ]);
-        if ($validatedRequest) {
-            $category = Category::find($request->id);
+
+        try {
+            $category = Category::find($id);
             $category->name = $request->name;
             $category->slug = $request->slug;
             $category->description = $request->description;
             $category->status = $request->status == null ? 0 : 1;
             $category->meta_title = $request->meta_title;
-            $category->meta_description = $request->meta_desc;
+            $category->meta_description = $request->meta_description;
+
             if ($request->hasFile('image')) {
                 $image = $request->file('image');
                 $imageName = "category" . '-' . $request->slug . '.' . $image->
@@ -79,24 +86,32 @@ class AdminCategoryController extends Controller
                 $image->move(public_path('images/categories'), $imageName);
                 $category->image = $imageName;
             }
+
             if ($category->isDirty() || $request->hasFile('image')) {
                 $category->save();
                 return redirect()->back();
             }
-
             return redirect()->back();
+        } catch (\Exception $e) {
+            return back()->withErrors(['message' => 'Something Wrong']);
         }
-
     }
 
     public function deleteCategory($id)
     {
         try {
             $category = Category::find($id);
-            $category->delete();
-            return 1;
+            if ($category) {
+                $oldFile = public_path('images/categories/' . $category->image);
+                if (file_exists($oldFile)) {
+                    unlink($oldFile);
+                }
+                $category->delete();
+                return redirect()->back();
+            }
         } catch (\Exception $e) {
-            return 0;
+            return redirect()->withErrors(['message' => 'Something Wrong!']);
         }
+
     }
 }
